+++ src/esdt/wipe-esdt.ts
@@ -8,7 +8,7 @@ import {
   TypedValue,
   Address,
   AddressValue,
-} from '@elrondnetwork/erdjs';
+} from '@multiversx/sdk-core';
 
 import { areYouSureAnswer, setup, commonTxOperations } from '../utils';
 import {