+++ src/meta-esdt/issue-meta-esdt.ts
@@ -9,7 +9,7 @@ import {
   ContractFunction,
   TypedValue,
   U32Value,
-} from '@elrondnetwork/erdjs';
+} from '@multiversx/sdk-core';
 
 import { areYouSureAnswer, setup, commonTxOperations } from '../utils';
 import {