+++ src/sft/send-sft.ts
@@ -5,7 +5,7 @@ import {
   Transaction,
   ESDTNFTTransferPayloadBuilder,
   Address,
-} from '@elrondnetwork/erdjs';
+} from '@multiversx/sdk-core';
 import axios from 'axios';
 
 import { areYouSureAnswer, setup, commonTxOperations } from '../utils';